feat(channel): Move isNotificationChannelSupported method to BuildUtil class.
